Refactor DI modules
- Refactor development server module
- Use staticmethods where possible
- Reuse injector in benchmark tests
