+++ dev/benchmark/get_statistics.py
@@ -1,11 +1,9 @@
 import random
 from decimal import Decimal
 
+from dev.benchmark.dependency_injection import benchmark_injector
 from tests.data_generators import PlanGenerator
 from tests.db.base_test_case import reset_test_db
-from tests.db.dependency_injection import DatabaseTestModule
-from tests.dependency_injection import TestingModule
-from workers_control.core.injector import Injector
 from workers_control.core.interactors import get_statistics
 from workers_control.core.records import ProductionCosts
 from workers_control.db.db import Database
@@ -13,7 +11,7 @@ from workers_control.db.db import Database
 
 class GetStatisticsBenchmark:
     def __init__(self) -> None:
-        self.injector = Injector([TestingModule(), DatabaseTestModule()])
+        self.injector = benchmark_injector
         reset_test_db()
         self.db = self.injector.get(Database)
         self.db.engine.dispose()
